add: border-radius to categorical progress bars
