+++ static_report/package.json
@@ -16,11 +16,6 @@
     "@emotion/styled": "^11.8.1",
     "@sentry/browser": "^7.8.0",
     "@sentry/tracing": "^7.8.0",
-    "@visx/axis": "^2.12.2",
-    "@visx/group": "^2.10.0",
-    "@visx/mock-data": "^2.1.2",
-    "@visx/scale": "^2.2.2",
-    "@visx/shape": "^2.12.2",
     "chart.js": "^3.9.1",
     "chartjs-plugin-datalabels": "^2.1.0",
     "d3": "^7.4.4",
