+++ static_report/src/components/shared/ColumnCard/ColumnTypeDetail/ColumnTypeDetailOther.tsx
@@ -7,7 +7,6 @@ import {
 import { getColumnDetails } from '../../../../utils/transformers';
 import { MetricCell } from '../../MetricCell';
 import { DataCompositionMetrics } from '../ColumnMetrics/DataCompositionMetrics';
-import { UniquenessMetrics } from '../ColumnMetrics/UniquenessMetrics';
 import { VALIDS } from './constants';
 
 type Props = { columnDatum: ColumnSchema };
